+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 # 0.1.2
 
+[_View the docs._](https://github.com/ukazap/permisi/blob/v0.1.2/README.md)
+
 - Fix namespaces/actions should no longer contain periods
 - Implement cache config for faster access to actor permissions
 
 # 0.1.1
+
+[_View the docs._](https://github.com/ukazap/permisi/blob/v0.1.1/README.md)
 
 - General code refactoring
 - Improvements on ActiveRecord backend:
@@ -14,6 +18,8 @@
 
 # 0.1.0
 
+[_View the docs._](https://github.com/ukazap/permisi/blob/v0.1.0/README.md)
+
 Finished extraction work from my past projects.
 
 - Implement ActiveRecord backend
@@ -21,5 +27,7 @@ Finished extraction work from my past projects.
 - Implement permissions hash sanitization and checking
 
 # 0.0.1
+
+[_View the docs._](https://github.com/ukazap/permisi/blob/v0.0.1/README.md)
 
 Reserved the gem name: https://en.wiktionary.org/wiki/permisi